requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.2
chromadb>=0.4.24
tiktoken>=0.6.0
numpy>=1.26.0
scikit-learn>=1.4.0
PyPDF2>=3.0.0
tabula-py>=2.9.0
pandas>=2.2.0
python-dotenv>=1.0.0
tqdm>=4.66.0
JPype1==1.4.1
langchain==0.1.9 